requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
